back/pyproject.toml
[project]
name = "task-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.38.32",
    "fastmcp>=2.7.0",
    "python-dotenv>=1.1.0",
]
